Use ffspart for more of PNOR assembly, notably ECC.
ffspart can automatically add ECC to partitions, and pflash can
add/remove it as needed based on the attribute of the partition. As
such, we can give ffspart the *non* ECC'd image to add.

We need to patch ffspart to accept ".ecc" files as already having ECC as
on POWER9 we add secure boot headers, which we don't *yet* construct
outside of update_image.pl/genPnorImages.pl. Accordingly, we can only
run the *first* "compare we got the same output" test on P8.

There are various rather odd things in PNOR assemlby, which we still
duplicate here (with notes that they're odd) in order to ensure we
produce an identical binary.

The next step would be to avoid update_image.pl completely for a P8 PNOR
image, followed by doing the signing in openpower-pnor rather than
hostboot scripts.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>